refactor: undo with memento snapshots
