--- src/atoms/actionLogAtom.ts
@@ -1,8 +1,0 @@
-import { atomWithImmer } from 'jotai/immer';
-
-interface ActionLogItem {
-  type: 'select' | 'unselect';
-  id: string;
-}
-
-export const actionLogAtom = atomWithImmer(new Array<ActionLogItem>());
