--- src/atoms/index.ts
@@ -1,2 +1,0 @@
-export { actionLogAtom } from './actionLogAtom';
-export { selectionAtom } from './selectionAtom';
