+++ src/routes/dashboard/atoms/mementoLogAtom.ts
@@ -1,0 +1,7 @@
+import { atomWithImmer } from 'jotai/immer';
+
+interface MementoEntry {
+  selection: Set<string>;
+}
+
+export const mementoLogAtom = atomWithImmer(new Array<MementoEntry>());
